requirements.txt
appdirs==1.4.4
attrs==24.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
capsolver_extension_python==1.0.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
chrome_extension_python==1.0.1
cssselect==1.2.0
curl_cffi==0.7.4
environs==11.2.1
fake-useragent==2.0.3
Faker==33.1.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
lxml==5.3.0
MailTm==0.0.81
marshmallow==3.23.2
mysql-connector-python==9.1.0
numpy==2.2.1
omkar_temp_mail==1.1.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parse==1.20.2
pycparser==2.22
pyee==11.1.1
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyVirtualDisplay==3.0
requests==2.32.3
selenium==4.27.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
undetected-chromedriver==3.5.5
urllib3==1.26.20
w3lib==2.2.1
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
zipp==3.21.0
